+++ src/compojure/api/middleware.clj
@@ -12,8 +12,6 @@
 
             [muuntaja.middleware]
             [muuntaja.core :as m]
-            [muuntaja.format.yaml :as yaml-format]
-            [muuntaja.format.msgpack :as msgpack-format]
 
             [ring.swagger.common :as rsc]
             [ring.swagger.middleware :as rsm]
@@ -126,14 +124,9 @@
   (or (:compojure.api.meta/serializable? response)
       (coll? (:body response))))
 
-(def muuntaja-options
-  (-> m/default-options
-      (yaml-format/with-yaml-format)
-      (msgpack-format/with-msgpack-format)))
-
 (defn create-muuntaja
   ([]
-   (create-muuntaja muuntaja-options))
+   (create-muuntaja m/default-options))
   ([muuntaja-or-options]
    (if muuntaja-or-options
      (if (instance? Muuntaja muuntaja-or-options)
@@ -188,7 +181,7 @@
 ;;
 
 (def api-middleware-defaults
-  {:formats muuntaja-options
+  {:formats m/default-options
    :exceptions {:handlers {::ex/request-validation ex/request-validation-handler
                            ::ex/request-parsing ex/request-parsing-handler
                            ::ex/response-validation ex/response-validation-handler